pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "mplgallery"
version = "0.1.0"
description = "Local-first CSV Plot Studio for drafting and editing Matplotlib plots from analysis tables."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Tanner" }
]
dependencies = [
  "streamlit>=1.35",
  "numpy>=1.26",
  "pandas>=2.0",
  "matplotlib>=3.8",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "typer>=0.12",
  "rich>=13.0",
  "pillow>=10.0"
]

[project.optional-dependencies]
dvc = ["dvc>=3.0"]
mlflow = ["mlflow>=2.0"]
integrations = ["dvc>=3.0", "mlflow>=2.0"]
desktop = ["pywebview>=5.4; platform_system == 'Windows'"]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
  "mypy>=1.8",
  "build>=1.2",
  "hatchling>=1.25"
]

[project.scripts]
mplgallery = "mplgallery.cli:main"

[project.gui-scripts]
mplgallery-desktop = "mplgallery.desktop:gui_main"

[tool.hatch.build.targets.wheel]
packages = ["src/mplgallery"]
exclude = [
  "src/mplgallery/ui/frontend/node_modules/**",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/mplgallery",
  "hatch_build.py",
  "README.md",
  "pyproject.toml",
]
exclude = [
  "src/mplgallery/ui/frontend/node_modules/**",
]

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
